src/test/java/calctest/CalculatorTest.java
package calctest;

import static org.testng.Assert.assertEquals;

import org.testng.annotations.Test;

import calc.Calculator;

public class CalculatorTest {
	Calculator calculator = new Calculator();

    @Test
    public void testAdd() {
        assertEquals(5, calculator.add(2, 3));
    }

    @Test
    public void testSubtract() {
        assertEquals(1, calculator.subtract(3, 2));
    }

    @Test
    public void testMultiply() {
        assertEquals(6, calculator.multiply(2, 3));
    }

    @Test
    public void testDivide() {
        assertEquals(2.5, calculator.divide(5, 2), 0.0001);
    }

   
}
